+++ .env.example
@@ -20,3 +20,5 @@ OKTA_CUSTOM_OAUTH_SERVER=
 OKTA_CLIENT_ID=
 OKTA_CLIENT_SECRET=
 SENTRY_DSN=
+REDIS_HOST=
+REDIS_PASSWORD=
