+++ src/server/models/Configuration.ts
@@ -25,6 +25,8 @@ export interface Configuration {
   aws: AWSConfiguration;
   githubRunNumber: string;
   sentryDsn: string;
+  rateLimiterConfiguration: RateLimiterConfiguration;
+  redisConfiguration: RedisConfiguration;
 }
 
 export interface AWSConfiguration {
@@ -64,4 +66,14 @@ export enum GU_API_DOMAIN {
   DEV = 'code.dev-guardianapis.com',
   CODE = 'code.dev-guardianapis.com',
   PROD = 'guardianapis.com',
+}
+
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface RateLimiterConfiguration {
+  //...
+}
+
+export interface RedisConfiguration {
+  password: string;
+  host: string;
 }
